up test coverage, remove dead code, fix dist
t/010_lookup_full.t will skip if the (none tracked) sample file is
missing as this sample file weighs in at over 400MB so i will not
add it to the dist

remove Readonly and just make the vars bog standard since we're not
really interested in enforcing them as constants (they're uppercase
which should be enough)

various tweaks in MANIFEST, Makefile.PL and reset Changes. update
perldoc and README.md. add travis and coveralls badges to POD